+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 - Update publish endpoint to the new Apps API
+- Update methods to deprecate, undeprecate an app
+- Create method to validate a release of an app
 
 ## [3.67.0] - 2019-12-03